Add api use in router
